gateway/requirements.txt
fastapi
uvicorn[standard]
requests
python-jose
grpcio
pydantic-settings